+++ R/shinyRatings.R
@@ -3,7 +3,7 @@
 #' @param inputId The input slot that will be used to access the value of number of stars.
 #' @param no_of_stars Number of stars that you want to display on the UI. 
 #' @param default Number of stars selected by default.
-#' 
+#' @param disabled Keep number of stars as fixed? Disable the option of changing values
 #' @return Ratings to be added to UI definition
 #' @examples 
 #' if(interactive()){
@@ -24,8 +24,8 @@
 #' 
 #' @export
 #'
-shinyRatings <- function(inputId, no_of_stars = 5, default = no_of_stars) {
-  do_checks(no_of_stars, default)
+shinyRatings <- function(inputId, no_of_stars = 5, default = no_of_stars, disabled = FALSE) {
+  do_checks(no_of_stars, default, disabled)
   # Since in HTML 2 check box constitute of 1 star
   calculate_def <- default * 2
   htmltools::tags$html(
@@ -36,14 +36,14 @@ shinyRatings <- function(inputId, no_of_stars = 5, default = no_of_stars) {
     ),
     htmltools::tags$body(
       shiny::div(id = inputId, class = "shinyRatings", 
-                 htmltools::HTML(ratings_html(no_of_stars)), data = jsonlite::toJSON(list(n = calculate_def))
+                 htmltools::HTML(ratings_html(no_of_stars, disabled)), data = jsonlite::toJSON(list(n = calculate_def))
                 )
     )
   )
 }
 
 #' @noRd
-ratings_html <- function(n) {
+ratings_html <- function(n, disabled) {
   half_num <- seq(0.5, n, 1)
   full_num <- seq(1, n, 1)
   half_num_chr <- sub('.', '', half_num, fixed = TRUE)
@@ -58,15 +58,21 @@ ratings_html <- function(n) {
       half_num, half_num_chr, half_num_chr, half_num, full_num, full_num_chr, full_num_chr, full_num
     ), collapse = '')
   
+  if(isTRUE(disabled)) {
+    # Add disabled attribute to input elements
+    dynamic_html <- gsub('<input', '<input disabled', dynamic_html, fixed = TRUE)
+  }
   sprintf('<div class="rating-group">%s</div>', dynamic_html)
 }
 
 #' @noRd
-do_checks <- function(no_of_stars, default) {
+do_checks <- function(no_of_stars, default, disabled) {
   # no_of_stars should be a whole number or should end with .5
   stopifnot("no_of_stars can be a whole number or a number ending with .5" = no_of_stars %% 1 %in% c(0, .5))
   # default value should be a whole number or should end with .5
   stopifnot("default value can be a whole number or a number ending with .5" = default %% 1 %in% c(0, .5))
   # default value should not be greater than no_of_stars
   stopifnot("default value can be greater than 0 and less than `no_of_stars`" = default > 0 && default <= no_of_stars)
+  # disabled can handle only logical values
+  stopifnot("disabled can take only logical values (TRUE/FALSE)" = is.logical(disabled))
 }